Add PrioritySort method to ServerAppService
A new private static method `PrioritySort` was added to the `ServerAppService` class to sort a list of `ServerAppInfo` objects based on a predefined priority list. The `PrioritySort` method first adds `ServerAppInfo` objects to the sorted list in the order specified by the priority list, then adds any remaining `ServerAppInfo` objects not in the priority list. The `ListAppsAsync` method was modified to call `PrioritySort` before returning the list, ensuring the list is sorted according to the specified priorities.